requirements.txt
flask==2.0.1
openai-whisper==20231117
torch==2.1.0 